dom/tests/mochitest/bugs/test_bug585819.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=585819
-->
<head>
  <title>Test for Bug 585819</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=585819">Mozilla Bug 585819</a>
<p id="display"></p>
<div id="content" style="display: none">
  <iframe></iframe>
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Bug 585819 **/

var range = document.createRange();
range.setStartBefore(document.body);
var fragment = range.createContextualFragment("<span></span>");
is(fragment.firstChild.localName, "span", "We don't want tag inference here!");

var iframeDoc = document.getElementsByTagName("iframe")[0].contentDocument;
var root = iframeDoc.documentElement;
is(root.localName, "html", "Wrong root.");
root.innerHTML = "<span></span>";
is(root.firstChild.localName, "head", "We want inference here!");

</script>
</pre>
</body>
</html>
